fix(frontend/friends): clean code and package added

+++ frontend/components/modals/friends/blockedsCont.tsx
@@ -8,10 +8,11 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 export function BlockedsCont() {
 	const { blockeds, revalid } = useAllBlocked();
 
-	async function unblock(e: any) {
+	async function unblock(e: any, userName: string) {
 		const { id } = e.currentTarget;
 		try {
-			await axios.delete(`blocked/${id}`);
+			if (window.confirm(`Do you wanna really unblock ${userName}?`))
+				await axios.delete(`blocked/${id}`);
 			window.alert(`user is unblocked`);
 			revalid();
 		} catch (e: any) {
@@ -21,11 +22,11 @@ export function BlockedsCont() {
 	return (
 		<div className="item-cont">
 			<ul className="d-flex column">
-				{blockeds?.map((el: any) => (
-					<li key={el.id} className="d-flex center gap p-2">
-						<Avatar url={el.image_url} size="sm" status={el.status} />
-						<span className="username">{el.name}</span>
-						<div id={el.id} onClick={unblock}>
+				{blockeds?.map((blocked: any) => (
+					<li key={blocked.id} className="d-flex center gap p-2">
+						<Avatar url={blocked.image_url} size="sm" status={blocked.status} />
+						<span className="username">{blocked.name}</span>
+						<div id={blocked.id} onClick={(e) => unblock(e, blocked.name)}>
 							<FontAwesomeIcon icon="lock-open" />
 						</div>
 					</li>
