+++ frontend/components/modals/friends/friendsCont.tsx
@@ -7,10 +7,11 @@ import { Loader } from "../../loader";
 export function FriendsCont() {
 	const { friends, revalid } = useAllFriend();
 
-	async function deleteFriend(e: any) {
+	async function deleteFriend(e: any, userName: string) {
 		const { id } = e.currentTarget;
 		try {
-			await axios.delete(`friend/${id}`);
+			if (window.confirm(`Do you really want delete ${userName}?`))
+				await axios.delete(`friend/${id}`);
 			window.alert(`user is deleted`);
 			revalid();
 		} catch (e: any) {
@@ -18,10 +19,11 @@ export function FriendsCont() {
 		}
 	}
 
-	async function blockFriend(e: any) {
+	async function blockFriend(e: any, userName: string) {
 		const { id } = e.currentTarget;
 		try {
-			await axios.post(`blocked`, { userId: id });
+			if (window.confirm(`Do you really want block ${userName}?`))
+				await axios.post(`blocked`, { userId: id });
 			window.alert(`user is blocked`);
 		} catch (e: any) {
 			window.alert(e.response.data.message);
@@ -33,15 +35,15 @@ export function FriendsCont() {
 	return (
 		<div className="item-cont">
 			<ul className="d-flex column">
-				{friends?.map((el: any) => (
-					<li key={el.id} className="d-flex center justify-between gap p-2">
-						<Avatar url={el.image_url} size="sm" status={el.status} />
-						<span className="username">{el.name}</span>
+				{friends?.map((friend: any) => (
+					<li key={friend.id} className="d-flex center justify-between gap p-2">
+						<Avatar url={friend.image_url} size="sm" status={friend.status} />
+						<span className="username">{friend.name}</span>
 						<div className="d-flex gap">
-							<div id={el.id} onClick={deleteFriend}>
+							<div id={friend.id} onClick={(e) => deleteFriend(e, friend.name)}>
 								<FontAwesomeIcon icon="user-minus" />
 							</div>
-							<div id={el.id} onClick={blockFriend}>
+							<div id={friend.id} onClick={(e) => blockFriend(e, friend.name)}>
 								<FontAwesomeIcon icon="ban" />
 							</div>
 						</div>
